Added support for resizing the window
The canvas used to take up a little bit more space than the size of the
window, this has been fixed using the added css code in page.html.

Resizing the window now resizes the canvas with it and redraws the
scene. As a result the canvas will always take up the whole window, and
the image will look as it's supposed to, even during/after a window
resize.