Finished lesson number 2
In this lesson we added color to our shapes.

The color of a pixel is determined by the fragment shader. In this case
the fragment shader receives the color from the vertex shader, which in
turn gets the color from our js code. In-between these two shaders the
colors undergo linear-interpolation, which creates a smooth gradient in
between two vertices.

In the javascript code we added support for the new attribute
'aVertexColor', and made sure it was passed to gl for the two shapes.

link to the lesson: http://learningwebgl.com/blog/?p=134