Finished lesson 3
In this lesson we animated the scene.

We defined a tick function, which gets initiated by the onLoad
function, and after that makes sure that it gets called repeatedly.

The animate function is used to update animation data (in this case
rotation of the shapes).

Lastly we created a mechanism to copy our current matrix, and restore
the old one when we're done. This allows us to apply transformations
relative to the previous transformations, without influencing drawings
that come after the current one.

Link to lesson: http://learningwebgl.com/blog/?p=239