Finished lesson 4
In this lesson we made the shapes actual 3D instead of 2d planes in a
3D world.

We're drawing the pyramid using 4 triangles and an array buffer
(nothing new here).

The cube is a bit more complex, we're using an element array buffer for
that one. This allows us to re-use vertices that we've specified
earlier in the positions buffer.

Link to lesson: http://learningwebgl.com/blog/?p=370